[REFACTOR] chatlist 페이지에서 formateDate을 use hook으로 빼기, 불필요한 useEffect 삭제
