+++ src/app/(root)/chat-list/page.tsx
@@ -4,18 +4,16 @@ import Page from '@/components/layout/Page';
 import { ChatStatusColor } from '@/components/message/ChatStatusColor';
 import { getChatList, subscribeChatList, untrackChatList } from '@/lib/api/SupabaseApi';
 import { supabase } from '@/lib/supabase-config';
-
 import { UserState, userState } from '@/recoil/user';
 import { ChatListType, LastMessageDataType, MessageType } from '@/types/realTimeType';
-import { format, formatDistanceToNow } from 'date-fns';
-import { ko } from 'date-fns/locale';
+
 import { useRouter } from 'next/navigation';
 import React, { useEffect, useState } from 'react';
 import { useRecoilState } from 'recoil';
 import useAlertModal from '@/components/common/modal/AlertModal';
 import { LastMessageArrayType } from '@/types/lastMessageArrayType';
-
 import useChatListModal from '@/components/common/modal/ChatListModal';
+import { formatDate } from '@/utils/useFormatDate';
 
 export default function ChatListPage() {
   const [chatList, setChatList] = useState<ChatListType[]>();
@@ -68,26 +66,8 @@ export default function ChatListPage() {
     };
   }, []);
 
-  useEffect(() => {
-    // console.log('lastMsg가 업데이트되었습니다');
-  }, [lastMsg]);
-
   // console.log('chatList:', chatList);
   // console.log('lastMessage:', lastMessageData);
-
-  const formatDate = (date: string) => {
-    const d = new Date(date);
-    const now = Date.now();
-    const diff = (now - d.getTime()) / 1000; // 현재 시간과의 차이(초)
-    if (diff < 60 * 1) {
-      return '방금 전';
-    }
-    if (diff < 60 * 60 * 24 * now) {
-      const distanceString = formatDistanceToNow(d, { addSuffix: true, locale: ko, includeSeconds: true });
-      return distanceString.replace(/^약\s*/, ''); // "약" 부분을 정규식을 사용하여 제거
-    }
-    return format(d, 'PPP EEE p', { locale: ko });
-  };
 
   const routerLink = (linkId: string, status: string) => {
     if (status === 'ACCEPT' || status === 'SOULMATE') {
